refactor(core): split parser.rs into tokenizer + scanner submodules
parser.rs (985 LOC) bundled two independent algorithms that share no
private state: the nom-combinator tokenizer (Token + parse_entity) and the
byte-level SIMD EntityScanner (+ the data_section_start HEADER-skip helper).

Carve them into parser/{tokenizer,scanner}.rs behind parser/mod.rs, which
re-exports Token, parse_entity, and EntityScanner so every consumer path
stays byte-identical: the internal crate::parser::* imports (streaming,
decoder, schema_gen) and the ifc_lite_core::{Token, parse_entity,
EntityScanner} root re-export (lib.rs) both resolve unchanged.

Pure code motion, no logic changes. All 14 parser unit tests travel with
their code (9 tokenizer, 5 scanner). cargo test -p ifc-lite-core green;
cargo build --workspace green; clippy clean.
